updating actions set up
